+++ src/ManyUI.jl
@@ -78,8 +78,10 @@ include("widgets/slider.jl")
 include("widgets/popup.jl")
 include("widgets/dropdown.jl")
 include("widgets/sparkline.jl")
+include("widgets/progresslist.jl")
 include("widgets/statusbar.jl")
 include("widgets/splitter.jl")
+include("widgets/dialog.jl")
 include("widgets/form.jl")
 include("precompile.jl")
 
@@ -195,6 +197,13 @@ export spark_bounds, spark_level
 
 # widgets/statusbar.jl
 export StatusBar, segments, status_layout
+
+# widgets/progresslist.jl
+export ProgressList, ProgressItem, n_items, set_progress!
+export pl_label_width, PL_GAP, PL_MIN_BAR
+
+# widgets/dialog.jl
+export Dialog, dialog_size, DIALOG_PAD, DIALOG_BUTTON_GAP
 
 # widgets/spinner.jl
 export Spinner